SCC: rename paragraph.py file to caption_paragraph.py
