+++ src/test/python/test_scc_paragraph.py
@@ -31,7 +31,7 @@ import unittest
 from fractions import Fraction
 
 from ttconv.model import ContentDocument, Span, Br
-from ttconv.scc.paragraph import SccCaptionParagraph
+from ttconv.scc.caption_paragraph import SccCaptionParagraph
 from ttconv.scc.style import SccCaptionStyle
 from ttconv.scc.caption_line import SccCaptionLine
 from ttconv.style_properties import TextAlignType
